move base-command to prevent it from appearing in help
