+++ packages/cli/src/commands/build.ts
@@ -1,5 +1,5 @@
 import { flags } from '@oclif/command'
-import BaseCommand from './base-command'
+import BaseCommand from '../base-command'
 import { compileProject } from '../services/config-service'
 import { checkCurrentDirIsABoosterProject } from '../services/project-checker'
 import { Script } from '../common/script'
