+++ packages/cli/src/commands/new/event-handler.ts
@@ -1,5 +1,5 @@
 import * as Oclif from '@oclif/command'
-import BaseCommand from '../base-command'
+import BaseCommand from '../../base-command'
 import {
   HasEvent,
   HasName,
