+++ packages/cli/src/commands/nuke.ts
@@ -1,5 +1,5 @@
 import { flags } from '@oclif/command'
-import BaseCommand from './base-command'
+import BaseCommand from '../base-command'
 import { nukeCloudProviderResources } from '../services/provider-service'
 import { compileProjectAndLoadConfig } from '../services/config-service'
 import { BoosterConfig, Logger } from '@boostercloud/framework-types'
